+++ src/components/passport-information-table/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./passport-information-table"
