+++ src/components/passport-information-table/passport-information-table.tsx
@@ -1,0 +1,69 @@
+import type { FC } from "react"
+import {
+  useReactTable,
+  getCoreRowModel,
+  getSortedRowModel,
+  flexRender,
+} from "@tanstack/react-table"
+import { Table, Thead, Tbody, Tr, Th, Td } from "@chakra-ui/react"
+import { TriangleDownIcon, TriangleUpIcon } from "@chakra-ui/icons"
+
+import type { PasseportInformationPerCountry } from "src/types"
+import { columns } from "./helpers"
+
+const PassportInformationTable: FC<{
+  data: PasseportInformationPerCountry[]
+}> = ({ data }) => {
+  const table = useReactTable({
+    columns,
+    data,
+    getCoreRowModel: getCoreRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+  })
+  return (
+    <Table>
+      <Thead>
+        {table.getHeaderGroups().map((headerGroup) => {
+          return (
+            <Tr key={headerGroup.id}>
+              {headerGroup.headers.map((header) => {
+                return (
+                  <Th
+                    key={header.id}
+                    onClick={header.column.getToggleSortingHandler()}
+                  >
+                    {header.isPlaceholder
+                      ? null
+                      : flexRender(
+                          header.column.columnDef.header,
+                          header.getContext()
+                        )}
+                    {header.column.getIsSorted() === "asc" && (
+                      <TriangleDownIcon />
+                    )}
+                    {header.column.getIsSorted() === "desc" && (
+                      <TriangleUpIcon />
+                    )}
+                  </Th>
+                )
+              })}
+            </Tr>
+          )
+        })}
+      </Thead>
+      <Tbody>
+        {table.getRowModel().rows.map((row) => (
+          <Tr key={row.id}>
+            {row.getVisibleCells().map((cell) => (
+              <Td key={cell.id}>
+                {flexRender(cell.column.columnDef.cell, cell.getContext())}
+              </Td>
+            ))}
+          </Tr>
+        ))}
+      </Tbody>
+    </Table>
+  )
+}
+
+export default PassportInformationTable
